cleaning the code and adding  Exception handling to the project

+++ LandBoundedContext/src/main/java/be/kdg/prog6/landSideBoundedContext/adapters/out/persistence/appointmentAndCalendar/AppointmentDataBaseAdapter.java
@@ -4,10 +4,15 @@ import be.kdg.prog6.landSideBoundedContext.adapters.out.entity.AppointmentEntity
 import be.kdg.prog6.landSideBoundedContext.adapters.out.entity.MaterialTypeEntity;
 import be.kdg.prog6.landSideBoundedContext.domain.Appointment;
 import be.kdg.prog6.landSideBoundedContext.port.in.ScheduleAppointmentUseCase;
+import be.kdg.prog6.landSideBoundedContext.util.errorClasses.AppointmentDontExistException;
+import be.kdg.prog6.landSideBoundedContext.util.errorClasses.CouldNotSaveAppointmentException;
+import be.kdg.prog6.landSideBoundedContext.util.errorClasses.DatabaseOperationException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.hibernate.exception.ConstraintViolationException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import be.kdg.prog6.landSideBoundedContext.port.out.AppointmentSavePort;
 @Service
@@ -24,21 +29,32 @@ public class AppointmentDataBaseAdapter implements AppointmentSavePort  {
     }
 
     @Override
-    public void saveAppointment(Appointment Appointment)  {
+    public void saveAppointment(Appointment appointment) {
         try {
-            appointmentRepository.save(modelMapper.map(Appointment, AppointmentEntity.class));
-        }
-        catch (Exception e) {
-            throw e;
+            appointmentRepository.save(modelMapper.map(appointment, AppointmentEntity.class));
+        } catch (ConstraintViolationException | DataAccessException e) {
+            throw new CouldNotSaveAppointmentException("Database error: Could not save the appointment. The appointment may already exist or there is a data integrity issue.");
+        } catch (Exception e) {
+            throw new CouldNotSaveAppointmentException("Unexpected error: Could not save the appointment.");
         }
     }
 
     @Override
-    public void deleteAppointment(Appointment Appointment)  {
-
-        AppointmentEntity appointmentEntity = appointmentRepository.findBySellerIdAndMaterialType(Appointment.getSellerId().id() , MaterialTypeEntity.valueOf(Appointment.getMaterialType().toString()));
-
-        appointmentRepository.delete(appointmentEntity);
+    public void deleteAppointment(Appointment appointment) {
+        AppointmentEntity appointmentEntity = appointmentRepository.findBySellerIdAndMaterialType(
+                appointment.getSellerId().id(),
+                MaterialTypeEntity.valueOf(appointment.getMaterialType().toString())
+        );
+        if (appointmentEntity == null) {
+            throw new AppointmentDontExistException("Appointment not found for seller ID: "
+                    + appointment.getSellerId().id() + " and material type: " + appointment.getMaterialType());
+        }
+        try {
+            appointmentRepository.delete(appointmentEntity);
+        } catch (DataAccessException e) {
+            throw new DatabaseOperationException("Failed to delete appointment due to database error " +  e);
+        }
     }
+
 
 }
